+++ src/modbus/resources/point/point_base.py
@@ -11,7 +11,7 @@ class ModbusPointBase(Resource):
     parser.add_argument('reg_length', type=int, required=True)
     parser.add_argument('type', type=str, required=True)
     parser.add_argument('enable', type=bool, required=True)
-    parser.add_argument('write_value', type=int, required=True)
+    parser.add_argument('write_value', type=float, required=True)
     parser.add_argument('data_type', type=str, required=True)
     parser.add_argument('data_endian', type=str, required=True)
     parser.add_argument('data_round', type=int, required=True)
